+++ docker.sh
@@ -10,9 +10,18 @@ export DEBUG=${DEBUG:-""}
 # Find out what branch we are on
 BRANCH=${BRANCH:-"$(git rev-parse --abbrev-ref HEAD)"}
 
+# Find out the version
+if [ "$BRANCH" = "master" ]; then
+    VERSION=""
+elif [ "${BRANCH}" = "develop" ]; then
+    VERSION="-dev"
+else
+    exit 0
+fi
+
 # Build individual images
-$DEBUG docker build -t hub.ncsa.illinois.edu/incore/data-jetty$VERSION:latest -f Dockerfile.data .
-$DEBUG docker build -t hub.ncsa.illinois.edu/incore/dfr3-jetty$VERSION:latest -f Dockerfile.dfr3 .
-$DEBUG docker build -t hub.ncsa.illinois.edu/incore/hazard-jetty$VERSION:latest -f Dockerfile.hazard .
-$DEBUG docker build -t hub.ncsa.illinois.edu/incore/space-jetty$VERSION:latest -f Dockerfile.space .
-$DEBUG docker build -t hub.ncsa.illinois.edu/incore/semantics-jetty$VERSION:latest -f Dockerfile.semantics .
+$DEBUG docker build -t hub.ncsa.illinois.edu/incore/data-jetty:latest$VERSION -f Dockerfile.data .
+$DEBUG docker build -t hub.ncsa.illinois.edu/incore/dfr3-jetty:latest$VERSION -f Dockerfile.dfr3 .
+$DEBUG docker build -t hub.ncsa.illinois.edu/incore/hazard-jetty:latest$VERSION -f Dockerfile.hazard .
+$DEBUG docker build -t hub.ncsa.illinois.edu/incore/space-jetty:latest$VERSION -f Dockerfile.space .
+$DEBUG docker build -t hub.ncsa.illinois.edu/incore/semantics-jetty:latest$VERSION -f Dockerfile.semantics .